requirements.txt
# Django Authentication Module Dependencies
# Install with: pip install -r requirements.txt

# Core Framework
Django>=4.0,<5.0

# REST API
djangorestframework>=3.13.0
django-filter>=22.0

# JWT Authentication
djangorestframework-simplejwt>=5.2.0

# API Documentation
drf-yasg>=1.21.0

# CORS Headers
django-cors-headers>=3.13.0

# Database
psycopg2-binary>=2.9.3  # PostgreSQL driver

# Environment Variables
python-dotenv>=0.21.0

# Image handling
Pillow>=9.2.0  # For image field support

# Email
django-anymail>=9.0  # For email backends (optional)

# Testing
pytest>=7.2.0
pytest-django>=4.5.0
pytest-cov>=4.0.0

# Code quality
flake8>=5.0.0
black>=22.12.0

# Production server
gunicorn>=20.1.0
whitenoise>=6.4.0

# Development
ipython>=8.0.0
django-debug-toolbar>=4.0.0

# Type hints
mypy>=0.991
django-stubs>=1.16.0

# Security
django-axes>=6.0.0  # Login attempt tracking (optional)
